+++ pkg/client/notificationMock_test.go
@@ -2,13 +2,12 @@ package client
 
 import (
 	"context"
-	"strings"
 	"testing"
 
-	"github.com/gesundheitscloud/go-svc/pkg/log"
-	"github.com/go-test/deep"
 	uuid "github.com/gofrs/uuid"
 	"github.com/stretchr/testify/assert"
+
+	"github.com/gesundheitscloud/go-svc/pkg/log"
 )
 
 func TestTestNotfifcationMock_GetNotifiedUsers(t *testing.T) {
@@ -107,76 +106,6 @@ func TestTestNotfifcationMock_GetNotifiedUsers(t *testing.T) {
 	}
 }
 
-func TestTestNotfifcationMock_LegacyGetNotifiedUsers(t *testing.T) {
-	// for the sake of the assertions, lets make the user IDs sorted
-	user1 := uuid.FromStringOrNil("22173f04-d443-4545-a883-680afd305141")
-	user2 := uuid.FromStringOrNil("461603ab-b71d-472f-8c1e-b965defdc6c7")
-	user3 := uuid.FromStringOrNil("519d75e2-61a5-44c2-93cd-476886cd5091")
-
-	type call struct {
-		templateKey string
-		language    string
-		subscribers []uuid.UUID
-	}
-	tests := []struct {
-		name  string
-		calls []call
-		want  NotifiedUsers
-	}{
-		{
-			"Initial empty state",
-			[]call{},
-			map[string]map[string][]uuid.UUID{},
-		},
-		{
-			"Single key two languages",
-			[]call{
-				{"key1", "en", []uuid.UUID{user1, user2}},
-				{"key1", "de", []uuid.UUID{user3}},
-			},
-			map[string]map[string][]uuid.UUID{
-				"key1": {
-					"en": []uuid.UUID{user1, user2},
-					"de": []uuid.UUID{user3},
-				},
-			},
-		},
-		{
-			"Two keys two languages",
-			[]call{
-				{"key1", "en", []uuid.UUID{user1, user2}},
-				{"key1", "de", []uuid.UUID{user3}},
-			},
-			map[string]map[string][]uuid.UUID{
-				"key1": {
-					"en": []uuid.UUID{user1, user2},
-					"de": []uuid.UUID{user3},
-				},
-			},
-		},
-		{
-			"Null conditions",
-			[]call{{"key1", "", []uuid.UUID{}}},
-			map[string]map[string][]uuid.UUID{
-				"key1": {
-					"": []uuid.UUID{},
-				},
-			},
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			c := NewNotificationMockLegacy()
-			for i := 0; i < len(tt.calls); i++ {
-				_ = c.SendTemplated(tt.calls[i].templateKey, tt.calls[i].language, nil, tt.calls[i].subscribers...)
-			}
-			got := c.GetNotifiedUsers()
-			assert.EqualValuesf(t, tt.want, got, "Notified users should match")
-		})
-	}
-}
-
 func TestNotificationMock_SendTemplated(t *testing.T) {
 	// for the sake of the assertions, lets make the user IDs sorted
 	user1 := uuid.FromStringOrNil("22173f04-d443-4545-a883-680afd305141")
@@ -215,13 +144,6 @@ func TestNotificationMock_SendTemplated(t *testing.T) {
 			wantStatus: NotificationStatus{
 				StateProcessing: "not ready yet",
 				StateQueue:      "not in queue",
-				ConsentStats: map[string]int{
-					EventConsent:          1,
-					EventRevoke:           1,
-					ConsentNeverConsented: 1,
-					ConsentUnknown:        0,
-					ConsentNotNeeded:      0,
-				},
 			},
 		},
 		{
@@ -240,13 +162,6 @@ func TestNotificationMock_SendTemplated(t *testing.T) {
 			wantStatus: NotificationStatus{
 				StateProcessing: "not ready yet",
 				StateQueue:      "not in queue",
-				ConsentStats: map[string]int{
-					EventConsent:          0,
-					EventRevoke:           0,
-					ConsentNeverConsented: 0,
-					ConsentUnknown:        0,
-					ConsentNotNeeded:      3,
-				},
 			},
 		},
 	}
@@ -268,11 +183,6 @@ func TestNotificationMock_SendTemplated(t *testing.T) {
 				assert.Equal(t, got.Error, tt.wantStatus.Error)
 				assert.Equal(t, got.Caller, tt.wantStatus.Caller)
 				assert.Equal(t, got.TraceID, tt.wantStatus.TraceID)
-				if diff := deep.Equal(tt.wantStatus.ConsentStats, got.ConsentStats); diff != nil {
-					t.Errorf("Arrays are not equal:\n%v", strings.Join(diff, "\n"))
-					t.Logf("Want: %+v", tt.wantStatus.ConsentStats)
-					t.Logf("Got: %+v", got.ConsentStats)
-				}
 			}
 		})
 	}